seo-manager-api/requirements.txt
# LLM + LangChain (let pip resolve compatible versions)
langchain>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0

# Environment variables
python-dotenv>=1.0.0

# Markdown -> HTML conversion
markdown>=3.9

# HTTP requests
requests>=2.32.0

# SerpAPI for competitive research
google-search-results>=2.4.2

# ChromaDB for semantic caching
chromadb>=0.4.0

# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
